feat: Testing gene ontology

+++ nolabs/features/protein_design/delete_experiment.py
@@ -1,5 +1,6 @@
 from nolabs.domain.experiment import ExperimentId
 from nolabs.features.protein_design.services.file_management import FileManagement
+from nolabs.exceptions import NoLabsException, ErrorCodes
 
 
 class DeleteExperimentFeature:
@@ -10,4 +11,8 @@ class DeleteExperimentFeature:
         assert id
 
         experiment_id = ExperimentId(id)
+
+        if not self._file_management.experiment_exists(experiment_id):
+            raise NoLabsException(message="Experiment does not exist", error_code=ErrorCodes.experiment_id_not_found)
+
         self._file_management.delete_experiment_folder(experiment_id)
